Fix logic for adding colors to squares.
When you change the size of the grid, the old grid is deleted and
replaced with a new grid.  The new grid squares were not changing colors
when hovering over them as it did with the default grid you start with.
This was because the eventListener originally added to the squares
was also deleted when the original grid was deleted.  I added code so
eventListeners would be added when the game is originally loaded and
when a new grid is rendered.